requirements.txt
apache-airflow
tensorflow
matplotlib
torch
torchvision